// src/shared/api/generated/types/FieldMetadata.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { FieldValue } from './FieldValue.ts'

export const fieldMetadataTypeEnum = {
  boolean: 'boolean',
  integer: 'integer',
  float: 'float',
  reference: 'reference',
} as const

export type FieldMetadataTypeEnum = (typeof fieldMetadataTypeEnum)[keyof typeof fieldMetadataTypeEnum]

/**
 * @description Метаданные поля
 * @example [object Object]
 */
export type FieldMetadata = {
  /**
   * @description Имя поля
   * @type string
   */
  name: string
  /**
   * @description Описание поля
   * @type string
   */
  description: string
  /**
   * @description Тип данных поля
   * @type string
   */
  type: FieldMetadataTypeEnum
  /**
   * @description Доступные операторы
   * @type array
   */
  operators: string[]
  /**
   * @description Доступные значения для поля
   */
  values?: FieldValue[] | null
}